2.3.0!

+++ src/manager/versionsManager.js
@@ -5,7 +5,7 @@ import LauncherManager from './launcherManager';
 import rimraf from 'rimraf';
 import ProfilesManager from './profilesManager';
 import LogManager from './logManager';
-const defaultVersion = require('../assets/defaultVersion.json');
+const defaultVersionForge = require('../assets/defaultVersion.json');
 const defaultVersionFabric = require('../assets/defaultVersionFabric.json');
 const version1710 = require('../assets/1710version.json');
 const semver = require('semver');
@@ -13,55 +13,54 @@ const VersionsManager = {
     getVersionsPath: function() {
         return path.join(Global.getMCPath(), '/versions')
     },
-    createVersion: function(profile, type) {
+    createVersion: function(profile, type, meta) {
         let versionname = `${profile.safename} [Minecraft Manager]`;
         if(!fs.existsSync(path.join(this.getVersionsPath(), versionname))) {
             fs.mkdirSync(path.join(this.getVersionsPath(), versionname));
         }
         let obj;
         if(type === 'forge') {
-            obj = defaultVersion;
+            obj = defaultVersionForge;
             if(this.checkIs1710OrLower(profile)) {
                 obj = version1710;
             }
             obj.id = versionname;
-            obj.inheritsFrom = profile.minecraftversion;
-            obj.jar = profile.minecraftversion;
-            obj.assets = profile.minecraftversion;
+            obj.inheritsFrom = profile.version.minecraft.version;
+            obj.jar = profile.version.minecraft.version;
+            obj.assets = profile.version.minecraft.version;
             obj.libraries[0].name = `minecraftmanager.profiles:mcm-${profile.id}:forge`;
+        }else if(type === 'forgeComplex') {
+            obj = meta.version;
+            obj['_comment_'] = undefined;
+            obj['time'] = undefined;
+            obj['releaseTime'] = undefined;
+            obj['libraries'][0] = {
+                "name": `minecraftmanager.profiles:mcm-${profile.id}:forge`
+            }
+            obj['id'] = versionname;
+            obj['jar'] = profile.version.minecraft.version;
+        }else if(type === 'fabric') {
+            obj = defaultVersionFabric;
+            obj.libraries = meta.launcherMeta.libraries.common;
+            obj.id = versionname;
+            obj.jar = profile.version.minecraft.version;
+            obj.inheritsFrom = profile.version.minecraft.version;
+    
+            obj.libraries.push({
+                name: `minecraftmanager.profiles:mcm-${profile.id}:fabric-intermediary`,
+                url: `https://maven.fabricmc.net`
+            });
+            obj.libraries.push({
+                name: `minecraftmanager.profiles:mcm-${profile.id}:fabric-loader`,
+                url: `https://maven.fabricmc.net`
+            });
         }
         fs.writeFile(path.join(this.getVersionsPath(), versionname, `${versionname}.json`), JSON.stringify(obj), () => {
-            profile.setVersion(versionname);
-            LauncherManager.setProfileData(profile, 'lastVersionId', versionname);
-        });
-    },
-    createVersionFabric: function(profile, meta) {
-
-        let versionname = `${profile.safename} [Minecraft Manager]`;
-        if(!fs.existsSync(path.join(this.getVersionsPath(), versionname))) {
-            fs.mkdirSync(path.join(this.getVersionsPath(), versionname));
-        }
-        let obj = defaultVersionFabric;
-        obj.libraries = meta.launcherMeta.libraries.common;
-        obj.id = versionname;
-        obj.jar = profile.minecraftversion;
-        obj.inheritsFrom = profile.minecraftversion;
-
-        obj.libraries.push({
-            name: `minecraftmanager.profiles:mcm-${profile.id}:fabric-intermediary`,
-            url: `https://maven.fabricmc.net`
-        });
-        obj.libraries.push({
-            name: `minecraftmanager.profiles:mcm-${profile.id}:fabric-loader`,
-            url: `https://maven.fabricmc.net`
-        });
-        fs.writeFile(path.join(this.getVersionsPath(), versionname, `${versionname}.json`), JSON.stringify(obj), () => {
-            profile.setVersion(versionname);
             LauncherManager.setProfileData(profile, 'lastVersionId', versionname);
         });
     },
     checkIs1710OrLower: function(profile) {
-        const ver = profile.minecraftversion;
+        const ver = profile.version.minecraft.version;
         switch(ver) {
             case '1.7.10':
                 return true;
@@ -80,7 +79,7 @@ const VersionsManager = {
         }
     },
     checkIs113OrHigher: function(profile) {
-        let version = profile.minecraftversion;
+        let version = profile.version.minecraft.version;
         if(version.split('.').length === 2) {
             let arr = version.split('.');
             arr.push('0');
@@ -88,45 +87,34 @@ const VersionsManager = {
         }
         return semver.gte(version, '1.13.0');
     },
-    renameVersion: function(profile, newName) {
+    renameVersion: function(profile, newName, type) {
         const oldVersionName = `${profile.safename} [Minecraft Manager]`;
         const newVersionName = `${newName} [Minecraft Manager]`;
 
         const oldVersionPath = path.join(this.getVersionsPath(), oldVersionName);
         const newVersionPath = path.join(this.getVersionsPath(), newVersionName);
 
-        let oldJSON = JSON.parse(fs.readFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`)));
-        oldJSON.id = newVersionName;
+        if(fs.existsSync(oldVersionPath)) {      
+            let oldJSON = JSON.parse(fs.readFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`)));
+            oldJSON.id = newVersionName;
 
-        // old library method
-        if(oldJSON.libraries[0].name.includes('minecraftmanager:profiles')) {
-            oldJSON.libraries[0].name = `minecraftmanager:profiles:mcm-${Global.createID(newName)}`;
-        }else{
-            oldJSON.libraries[0].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:forge`;
+            if(type === 'forge' || type === 'forgeComplex') {
+                // old library method
+                if(oldJSON.libraries[0].name.includes('minecraftmanager:profiles')) {
+                    oldJSON.libraries[0].name = `minecraftmanager:profiles:mcm-${Global.createID(newName)}`;
+                }else{
+                    oldJSON.libraries[0].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:forge`;
+                }
+            }else if(type === 'fabric') {
+                oldJSON.libraries[oldJSON.libraries.length - 1].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:fabric-loader`;
+                oldJSON.libraries[oldJSON.libraries.length - 2].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:fabric-intermediary`;    
+            }
+
+            fs.writeFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`), JSON.stringify(oldJSON));
+
+            fs.renameSync(path.join(oldVersionPath, `/${oldVersionName}.json`), path.join(oldVersionPath, `/${newVersionName}.json`));
+            fs.renameSync(oldVersionPath, newVersionPath);
         }
-
-        fs.writeFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`), JSON.stringify(oldJSON));
-
-        fs.renameSync(path.join(oldVersionPath, `/${oldVersionName}.json`), path.join(oldVersionPath, `/${newVersionName}.json`));
-        fs.renameSync(oldVersionPath, newVersionPath);
-    },
-    renameVersionFabric: function(profile, newName) {
-        const oldVersionName = `${profile.safename} [Minecraft Manager]`;
-        const newVersionName = `${newName} [Minecraft Manager]`;
-
-        const oldVersionPath = path.join(this.getVersionsPath(), oldVersionName);
-        const newVersionPath = path.join(this.getVersionsPath(), newVersionName);
-
-        let oldJSON = JSON.parse(fs.readFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`)));
-        oldJSON.id = newVersionName;
-
-        oldJSON.libraries[oldJSON.libraries.length - 1].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:fabric-loader`;
-        oldJSON.libraries[oldJSON.libraries.length - 2].name = `minecraftmanager.profiles:mcm-${Global.createID(newName)}:fabric-intermediary`;
-
-        fs.writeFileSync(path.join(oldVersionPath, `/${oldVersionName}.json`), JSON.stringify(oldJSON));
-
-        fs.renameSync(path.join(oldVersionPath, `/${oldVersionName}.json`), path.join(oldVersionPath, `/${newVersionName}.json`));
-        fs.renameSync(oldVersionPath, newVersionPath);
     },
     deleteVersion: function(profile) {
         return new Promise((resolve) => {
